+++ src/Concerto/TestBundle/Resources/R/service.R
@@ -1,0 +1,7 @@
+require(concerto5)
+
+concerto.log("starting service listener")
+
+
+
+concerto.log("listener service closing")