+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.session.serialize.R
@@ -1,13 +1,13 @@
 concerto.session.serialize <- function(){
     concerto.log("serializing session...")
 
-    serialized = serialize(concerto, NULL)
-
     if(concerto$sessionStorage == "redis") {
+        #TODO add comppression
+        serialized = serialize(concerto, NULL)
         expSeconds = as.numeric(concerto$sessionFilesExpiration) * 24 * 60 * 60
         concerto$redisConnection$SETEX(concerto$session$hash, expSeconds, serialized)
     } else {
-        writeBin(serialized, concerto$sessionFile)
+        save(concerto, file=concerto$sessionFile)
     }
 
     concerto.log("session serialized")