+++ cron/concerto.forker.guard.sh
@@ -4,7 +4,7 @@
 
 PROCESS="forker.R";
 
-if ! ps ax | grep -v grep | grep $PROCESS > /dev/null; then
+if ! ps --ppid 1 -F | grep -v grep | grep $PROCESS > /dev/null; then
   echo "`date`: forker process not running";
   /usr/bin/php /app/concerto/bin/console concerto:forker:start --env=prod;
 fi